requirements.txt
python-dotenv>=1.0.1
faiss-cpu>=1.8.0
numpy>=1.24.0
sentence-transformers>=3.0.0
pymupdf>=1.24.0
google-generativeai>=0.7.2
